Wired up Django Settings. Set up Docker with Envoy.
I went ahead and set things up with envoy because I already have some personal
templates to make that work out. I'm pushing these changes mainly to preserve
the work, as I'm not confidant they would operate smoothly out of the box.
Looking forward to adding the docker layer to SpaceMacs in the hopes that it
will smooth out development time. On that note, I need to look into some process
that will enable hot reloading for the envoy image I'm bulding the django
project on top of. Really, I need to learn multi-stage builds, if only to reduce
the ***attack surface area!!***, which is a legitmate concern once production
hits.