Witness the swidden.
As the branchname suggests, the first growth needed burning! I've been in the
bad habit of overengineering out of the gate and got bit this go round. Between
a recent switch to zsh, a somewhat overzealous destruction of unwanted dotfiles,
and probably many other own goals, the dust has settled in my development
environment. Given that I'm not working with anyone on this project and have the
luxury of containing things to a single computer, I'll let my environment handle
things and avoid docker development, thankfully! That should vastly reduce time
 spent rebuilding images and allow me to focus on what matters. I'll likely
 merge swidden into master sooner or later, likely sooner.