Wired in Basic JWT Authentication.
Tokens will be passed for recognized users, but I haven't arranged for a token
to be created during user creation. Honestly, I'm probably going to be lazy on
that front and table tokens at creation in favor for a system where users have
to log in after account creation. If this is positively awful from a user
standpoint, we can make the necessary changes after getting feedback.