Built up Basic Serialization, Established URL Patterns.
Yay, that's done. Just a bunch of boilerplate. Now, for the real task of
enabling serializers to filter based on JSON requests. However, I think this is
a good opportunity to do to things...

* Explore restclient as an alternative to PostMan for endpoint testing.
* Start building out React frontend.

So yeah, this is a nice enough inflection point. Hopefully some actual progress
is in the offing!