Well, We All Have to Learn Some Day.
Here I am starting my project without a proper gitignore, and therefore having
to undergo the tedious task fo untracking a variety of files that shouldn't have
been indexed in the first place. Definitely need to be more cognizant of these
things earlier in the process. Thankfully it's not so late though...